add up down simulation
