+++ Assets/Scenes/OneLegModel.unity
@@ -452,7 +452,7 @@ HingeJoint:
   m_GameObject: {fileID: 1003538082850180216}
   m_ConnectedBody: {fileID: 3042904687054863569}
   m_Anchor: {x: 25, y: 0, z: -384.5}
-  m_Axis: {x: -1, y: 0, z: 0}
+  m_Axis: {x: 1, y: 0, z: 0}
   m_AutoConfigureConnectedAnchor: 1
   m_ConnectedAnchor: {x: 24.999952, y: -0.000023841858, z: -384.50003}
   m_UseSpring: 0
@@ -640,7 +640,7 @@ HingeJoint:
     targetPosition: 0
   m_UseMotor: 0
   m_Motor:
-    targetVelocity: 100
+    targetVelocity: -100
     force: 50
     freeSpin: 0
   m_UseLimits: 0
