+++ Assets/Scripts/sample_leg_up_down.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 131dce66518de49a5b98f09dc53e8d0f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
